refactor(functional-tests): distribute functional tests across workers evenly (part 2) Because:
* we want to execute the functional tests in CI as fast as possible

This commit:

* matches parallelism to the number of cores for a resource_class (4 for large and 8 for xlarge)
* splits and refactors the misc/keyStretchingV2.spec module into:
  * key-stretching-v2/totp.spec.ts
  * key-stretching-v2/signinBlocked.spec.ts
* splits the oauth/oauthPermissions.spec.ts module into:
  * oauth/oauthPermissionsSignin.spec.ts
  * oauth/oauthPermissionsSignup.spec.ts
* splits part of the oauth/signin.spec.ts module into oauth/signinBlocked.spec.ts
* splits part of the react-conversion/oauthResetPassword.spec.ts module into:
  * react-conversion/oauthResetPasswordRecoveryKey.spec.ts
  * react-conversion/oauthResetPasswordScopeKey.spec.ts
  * react-conversion/oauthResetPasswordSyncMobile.spec.ts
* splits part of the react-conversion/signup.spec.ts module into react-conversion/oauthSignup.spec.ts
* splits part of the settings/changeEmail.spec.ts module into settings/changeEmailBlocked.spec.ts
* splits part of the settings/changePassword.spec.ts module into settings/changePasswordValidation.spec.ts
* splits part of the subscription-tests/coupons-ui-tests/coupon.spec.ts module into:
  * subscription-tests/coupons-ui-tests/couponExpired.spec.ts
  * subscription-tests/coupons-ui-tests/couponForeverDiscount.spec.ts
  * subscription-tests/coupons-ui-tests/couponInvalid.spec.ts
  * subscription-tests/coupons-ui-tests/couponOneTimeDiscount.spec.ts
* splits part of the subscription-tests/coupons-ui-tests/resubscription.spec.ts module into subscription-tests/coupons-ui-tests/paymentMode.spec.ts
* splits part of the syncV3/fxDesktopHandshake.spec.ts module into syncV3/fxDesktopHandshakeNonSync.spec.ts
* splits part of the syncV3/fxDesktopV3ForceAuth.spec.ts module into syncV3/fxDesktopV3ForceAuthUnregistered.spec.ts
* splits part of the syncV3/signIn.spec.ts module into syncV3/oauthSignIn.spec.ts
* fixes confirmResetPasswordHeading error in react-conversion/resetPassword.spec.ts

Closes #FXA-9699
