+++ packages/functional-tests/pages/signinTotpCode.ts
@@ -9,6 +9,8 @@ export class SigninTotpCodePage extends BaseTokenCodePage {
 
   get input() {
     this.checkPath();
-    return this.page.locator('input[type=number]');
+    return this.page
+      .getByTestId('signin-totp-code-input-label') // React
+      .or(this.page.getByPlaceholder('Enter 6-digit code')); //Backbone
   }
 }
