+++ packages/functional-tests/tests/settings/changePassword.spec.ts
@@ -69,70 +69,6 @@ test.describe('severity-1 #smoke', () => {
       await expect(settings.primaryEmail.status).toHaveText(credentials.email);
     });
 
-    const testCases = [
-      {
-        name: 'short password',
-        error: 'At least 8 characters',
-        password: '2short',
-        confirmPassword: '2short',
-      },
-      {
-        name: 'common password',
-        error: 'Not a commonly used password',
-        password: 'passwords',
-        confirmPassword: 'passwords',
-      },
-      {
-        name: 'mismatch confirm password',
-        error: 'New password matches confirmation',
-        password: 'newPasswordTest',
-        confirmPassword: 'newPasswordTest2',
-      },
-    ];
-    for (const { name, error, password, confirmPassword } of testCases) {
-      test(`new password validation - ${name}`, async ({
-        target,
-        pages: { page, login, settings, changePassword },
-        testAccountTracker,
-      }) => {
-        await signInAccount(target, page, login, testAccountTracker);
-
-        await settings.goto();
-        await settings.password.changeButton.click();
-
-        await expect(changePassword.changePasswordHeading).toBeVisible();
-
-        await changePassword.newPasswordTextbox.fill(password);
-        await changePassword.confirmPasswordTextbox.fill(confirmPassword);
-
-        await expect(changePassword.passwordError).toHaveText(error);
-      });
-    }
-
-    test(`new password validation - email as password`, async ({
-      target,
-      pages: { page, login, settings, changePassword },
-      testAccountTracker,
-    }) => {
-      const { email } = await signInAccount(
-        target,
-        page,
-        login,
-        testAccountTracker
-      );
-
-      await settings.goto();
-      await settings.password.changeButton.click();
-
-      await expect(changePassword.changePasswordHeading).toBeVisible();
-
-      await changePassword.newPasswordTextbox.fill(email);
-
-      await expect(changePassword.passwordError).toHaveText(
-        'Not your email address'
-      );
-    });
-
     test('change password with short password tooltip shows, cancel and try to change password again, tooltip is not shown', async ({
       target,
       pages: { page, login, settings, changePassword },
