+++ packages/functional-tests/tests/subscription-tests/coupons-ui-tests/couponOneTimeDiscount.spec.ts
@@ -1,0 +1,107 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import { Page, expect, test } from '../../../lib/fixtures/standard';
+import { BaseTarget, Credentials } from '../../../lib/targets/base';
+import { TestAccountTracker } from '../../../lib/testAccountTracker';
+import { LoginPage } from '../../../pages/login';
+
+test.describe('severity-2 #smoke', () => {
+  test.describe('coupon test one time discount', () => {
+    test.beforeEach(() => {
+      test.slow();
+    });
+
+    test('subscribe with a one time discount coupon', async ({
+      target,
+      page,
+      pages: { relier, subscribe, login },
+      testAccountTracker,
+    }, { project }) => {
+      test.skip(
+        project.name === 'production',
+        'no real payment method available in prod'
+      );
+      await signInAccount(target, page, login, testAccountTracker);
+
+      await relier.goto();
+      await relier.clickSubscribe12Month();
+
+      // 'auto50ponetime' is a one time 50% discount coupon for a 12mo plan
+      await subscribe.addCouponCode('auto50ponetime');
+
+      // Verify the coupon is applied successfully
+      await expect(subscribe.promoCodeAppliedHeading).toBeVisible();
+      expect(await subscribe.oneTimeDiscountSuccess()).toBe(true);
+
+      // Verify the line items is visible after applying discount
+      expect(await subscribe.discountListPrice()).toBe(true);
+      expect(await subscribe.discountLineItem()).toBe(true);
+
+      // Successfully subscribe
+      await subscribe.setConfirmPaymentCheckbox();
+      await subscribe.setFullName();
+      await subscribe.setCreditCardInfo();
+      await subscribe.clickPayNow();
+      await subscribe.submit();
+      await relier.goto();
+      await relier.clickEmailFirst();
+      await login.submit();
+      expect(await relier.isPro()).toBe(true);
+    });
+
+    test('remove a coupon and verify', async ({
+      pages: { relier, subscribe, login },
+    }, { project }) => {
+      test.fixme(
+        project.name !== 'local',
+        'Fix required as of 2024/05/13 (see FXA-9689).'
+      );
+      test.skip(
+        project.name === 'production',
+        'test plan not available in prod'
+      );
+
+      await relier.goto();
+      await relier.clickSubscribe12Month();
+
+      // 'auto50ponetime' is a one time 50% discount coupon for a 12mo plan
+      await subscribe.addCouponCode('auto50ponetime');
+
+      // Verify the coupon is applied successfully
+      await expect(
+        await subscribe.getCouponStatusByDataTestId('coupon-remove-button')
+      ).toBeVisible({
+        timeout: 5000,
+      });
+      expect(await subscribe.oneTimeDiscountSuccess()).toBe(true);
+
+      // Verify the line items is visible after applying discount
+      expect(await subscribe.discountListPrice()).toBe(true);
+      expect(await subscribe.discountLineItem()).toBe(true);
+
+      // Remove the coupon
+      await subscribe.removeCouponCode();
+
+      // Verify the discount is removed
+      expect(await subscribe.discountLineItem()).toBe(false);
+    });
+  });
+});
+
+async function signInAccount(
+  target: BaseTarget,
+  page: Page,
+  login: LoginPage,
+  testAccountTracker: TestAccountTracker
+): Promise<Credentials> {
+  const credentials = await testAccountTracker.signUp();
+  await page.goto(target.contentServerUrl);
+  await login.fillOutEmailFirstSignIn(credentials.email, credentials.password);
+
+  //Verify logged in on Settings page
+  expect(await login.isUserLoggedIn()).toBe(true);
+
+  return credentials;
+}
