+++ packages/functional-tests/tests/syncV3/fxDesktopV3ForceAuth.spec.ts
@@ -89,99 +89,6 @@ test.describe('severity-1 #smoke', () => {
       await fxDesktopV3ForceAuth.checkWebChannelMessage('fxaccounts:login');
     });
 
-    test('sync v3 with an unregistered email, no uid', async ({
-      pages: { configPage },
-      syncBrowserPages: { fxDesktopV3ForceAuth, login },
-      testAccountTracker,
-    }) => {
-      const config = await configPage.getConfig();
-      test.skip(
-        config.showReactApp.signUpRoutes === true,
-        'force_auth is no longer supported for signup with react, FXA-9410'
-      );
-
-      const credentials = await testAccountTracker.signUpSync();
-      const email = testAccountTracker.generateSyncEmail();
-
-      await fxDesktopV3ForceAuth.openWithReplacementParams(credentials, {
-        email,
-        uid: undefined,
-      });
-      const error = await login.signInError();
-      expect(error).toContain('Recreate');
-      const emailInputValue = await login.getEmailInput();
-      expect(emailInputValue).toBe(email);
-      const emailInput = await login.getEmailInputElement();
-      await expect(emailInput).toBeDisabled();
-      expect(await (await login.getUseDifferentAccountLink()).count()).toEqual(
-        0
-      );
-      await login.fillOutFirstSignUp(email, credentials.password, {
-        enterEmail: false,
-      });
-      await fxDesktopV3ForceAuth.checkWebChannelMessage(
-        'fxaccounts:can_link_account'
-      );
-      await fxDesktopV3ForceAuth.checkWebChannelMessage('fxaccounts:login');
-    });
-
-    test('sync v3 with an unregistered email, registered uid', async ({
-      pages: { configPage },
-      syncBrowserPages: { fxDesktopV3ForceAuth, login },
-      testAccountTracker,
-    }) => {
-      const config = await configPage.getConfig();
-      test.skip(
-        config.showReactApp.signUpRoutes === true,
-        'force_auth is no longer supported for signup with react, FXA-9410'
-      );
-
-      const credentials = await testAccountTracker.signUpSync();
-      const email = testAccountTracker.generateSyncEmail();
-
-      await fxDesktopV3ForceAuth.openWithReplacementParams(credentials, {
-        email,
-      });
-      const error = await login.signInError();
-      expect(error).toContain('Recreate');
-      const emailInputValue = await login.getEmailInput();
-      expect(emailInputValue).toBe(email);
-      const emailInput = await login.getEmailInputElement();
-      await expect(emailInput).toBeDisabled();
-      expect(await (await login.getUseDifferentAccountLink()).count()).toEqual(
-        0
-      );
-    });
-
-    test('sync v3 with an unregistered email, unregistered uid', async ({
-      pages: { configPage },
-      syncBrowserPages: { fxDesktopV3ForceAuth, login },
-      testAccountTracker,
-    }) => {
-      const config = await configPage.getConfig();
-      test.skip(
-        config.showReactApp.signUpRoutes === true,
-        'force_auth is no longer supported for signup with react, FXA-9410'
-      );
-
-      const credentials = await testAccountTracker.signUpSync();
-      const email = testAccountTracker.generateSyncEmail();
-      const uid = makeUid();
-      await fxDesktopV3ForceAuth.openWithReplacementParams(credentials, {
-        email,
-        uid,
-      });
-      const error = await login.signInError();
-      expect(error).toContain('Recreate');
-      const emailInputValue = await login.getEmailInput();
-      expect(emailInputValue).toBe(email);
-      const emailInput = await login.getEmailInputElement();
-      await expect(emailInput).toBeDisabled();
-      expect(await (await login.getUseDifferentAccountLink()).count()).toEqual(
-        0
-      );
-    });
-
     test('blocked with an registered email, unregistered uid', async ({
       syncBrowserPages: {
         page,
